move support bundles to services
